+++ admin/js/controllers-admin/backgrounds.js
@@ -17,7 +17,6 @@ postworldAdmin.directive( 'pwAdminBackgrounds', [ function(){
 }]);
 
 postworldAdmin.controller('pwAdminBackgroundsCtrl',
-	[ '$scope', '$log', '$window', '$parse', '$pwData', '$_', 'pwPostOptions', '$pw', 
 	function ( $scope, $log, $window, $parse, $pwData, $_, $pwPostOptions, $pw ) {
 	
 	$scope.view = 'settings';
@@ -80,4 +79,4 @@ postworldAdmin.controller('pwAdminBackgroundsCtrl',
 
 
 	
-}]);
+});