+++ admin/php/page-feeds.php
@@ -35,7 +35,6 @@
 ?>
 <script>
 	postworldAdmin.controller( 'pwFeedsDataCtrl',
-		[ '$scope', '$_', '$timeout', 'pwPostOptions',
 		function( $scope, $_, $timeout, $pwPostOptions ){
 		$scope.pwFeeds = <?php echo json_encode( $pwFeeds ); ?>;
 		$scope.pwFeedSettings = <?php echo json_encode( $pwFeedSettings ); ?>;
@@ -80,7 +79,7 @@
 				);
 		}
 
-	}]);
+	});
 </script>
 
 <div class="postworld feeds wrap" ng-cloak>