+++ php/widgets/related-posts/related-posts-admin.php
@@ -28,7 +28,7 @@
 
 	///// CONTROLLER /////
 	postworldAdmin.controller('<?php echo $instance ?>Ctrl',
-		['$scope', 'pwData', '_', '$log',
+		['$scope', '$pwData', '$_', '$log',
 		function( $scope, $pwData, $_, $log ) {
 			$log.debug( 'Init Related Posts Widget Controller : ' + '<?php echo $instance ?>' );
 			$scope.settings = <?php echo json_encode( $options['settings'] ); ?>;