+++ templates/galleries/gallery-inline.php
@@ -1,16 +1,15 @@
 <!-- POSTWORLD GALLERY SHORTCODE : <?php echo $vars['instance']; ?> -->
 <script>
 	postworld.controller( '<?php echo $vars['instance']; ?>',
-		[ '$scope', '$_', '$pw', 'pwImages', 'pwPosts', 'pwModal','$pwData',
 		function( $scope, $_, $pw, $pwImages, $pwPosts, $pwModal, $pwData ){
 		var instance = "<?php echo $vars['instance']; ?>";
 		var galleryInstance = "gallery-" + instance;
 		var galleryPosts = <?php echo json_encode( $vars['posts'] ); ?>;
-
+		
 		$pwData.insertFeed( galleryInstance, { posts: galleryPosts } );
 		$scope.feed = $pwPosts.getFeed( galleryInstance );
 
-	}]);
+	});
 	pwRegisterController( "<?php echo $vars['instance']; ?>" );
 </script>